refactor: add share button group, modify some text, hidden developer
